python/requirements.txt
hashlib
io
pandas
requests
bs4
pillow
selenium
Pyarrow
validators